requirements.txt
kafka-python
pandas
librosa
psycopg2-binary
soundfile
